♻️ Group together flow and fuzz test executor and add generic client with fuzz client bound
