+++ crates/client/tests/fuzz_template/test_fuzz.rs
@@ -7,12 +7,15 @@ mod types;
 use additional_program::entry as entry_additional_program;
 use idl_test::entry as entry_idl_test;
 pub use transactions::*;
-#[derive(Default, FuzzTestExecutor)]
-struct FuzzTest {
-    client: TridentSVM,
+#[derive(Default)]
+struct FuzzTest<C> {
+    client: C,
 }
 #[flow_executor]
-impl FuzzTest {
+impl<C: FuzzClient + std::panic::RefUnwindSafe> FuzzTest<C> {
+    fn new(client: C) -> Self {
+        Self { client }
+    }
     #[init]
     fn start(&mut self) {
         self.client.deploy_native_program(ProgramEntrypoint::new(
