--- crates/fuzz/derive/fuzz_test_executor/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "trident-derive-fuzz-test-executor"
-version = "0.0.5"
-rust-version = "1.60"
-edition = "2021"
-license-file = "../../../../LICENSE"
-readme = "../../../../README.md"
-description = "trident-derive-fuzz-test-executor"
-
-[lib]
-proc-macro = true
-doctest = false
-
-[dependencies]
-proc-macro2 = "1.0"
-quote = "1.0"
-syn = { version = "2.0", features = ["full"] }
-trident-syn = { path = "../../../syn" }
