+++ crates/syn/src/codegen/mod.rs
@@ -1,6 +1,5 @@
 pub mod trident_accounts;
 pub mod trident_flow_executor;
-pub mod trident_fuzz_test_executor;
 pub mod trident_instruction;
 pub mod trident_remaining_accounts;
 pub mod trident_transaction;
