+++ crates/syn/src/codegen/trident_flow_executor.rs
@@ -60,6 +60,37 @@ impl ToTokens for TridentFlowExecutorImpl {
                     }
                     Ok(())
                 }
+
+                fn fuzz(&mut self) {
+                    if cfg!(honggfuzz) {
+                        loop {
+                            fuzz_honggfuzz(|fuzzer_data| {
+                                let mut buf = Unstructured::new(fuzzer_data);
+                                let _ = self.execute_flows(&mut buf);
+                                self.client.clear_accounts();
+                            });
+                        }
+                    } else if cfg!(afl) {
+                        fuzz_afl(true, |fuzzer_data| {
+                            let mut buf = Unstructured::new(fuzzer_data);
+                            let _ = self.execute_flows(&mut buf);
+                            self.client.clear_accounts();
+                        });
+                    } else if cfg!(honggfuzz_debug) {
+                        let mut crash_file = String::new();
+                        std::io::stdin()
+                            .read_line(&mut crash_file)
+                            .expect("Failed to read crash file path from stdin");
+                        let crash_file = crash_file.trim();
+
+                        let fuzzer_data = std::fs::read(crash_file).expect("Failed to read crash file");
+                        let mut buf = Unstructured::new(&fuzzer_data);
+                        let _ = self.execute_flows(&mut buf);
+                        self.client.clear_accounts();
+                    } else {
+                        panic!("Select Honggfuzz or AFL for fuzzing!!!")
+                    }
+                }
             }
         };
