+++ examples/cpi/cpi-metaplex-7/trident-tests/fuzz_0/test_fuzz.rs
@@ -6,12 +6,15 @@ mod transactions;
 mod types;
 use cpi_metaplex_7::entry as entry_cpi_metaplex_7;
 pub use transactions::*;
-#[derive(Default, FuzzTestExecutor)]
-struct FuzzTest {
-    client: TridentSVM,
+#[derive(Default)]
+struct FuzzTest<C> {
+    client: C,
 }
 #[flow_executor]
-impl FuzzTest {
+impl<C: FuzzClient + std::panic::RefUnwindSafe> FuzzTest<C> {
+    fn new(client: C) -> Self {
+        Self { client }
+    }
     #[init]
     fn start(&mut self) {
         self.client.deploy_native_program(ProgramEntrypoint::new(
