+++ examples/hello_world/trident-tests/fuzz_0/test_fuzz.rs
@@ -6,12 +6,17 @@ mod transactions;
 mod types;
 use hello_world::entry as entry_hello_world;
 pub use transactions::*;
-#[derive(Default, FuzzTestExecutor)]
-struct FuzzTest {
-    client: TridentSVM,
+
+#[derive(Default)]
+struct FuzzTest<C> {
+    client: C,
 }
+
 #[flow_executor]
-impl FuzzTest {
+impl<C: FuzzClient + std::panic::RefUnwindSafe> FuzzTest<C> {
+    fn new(client: C) -> Self {
+        Self { client }
+    }
     #[init]
     fn start(&mut self) {
         self.client.deploy_native_program(ProgramEntrypoint::new(
